Правка характеристик
Правил типы характеристик(была недоработка в том что те значения которые должны быть целочисленными были флоатами), а также добавил регуляцию кд способности и ее время действия через улучшения

+++ Assets/Scripts/ScriptableObjectsExamples/Weapon.cs
@@ -7,7 +7,7 @@ public class Weapon : ScriptableObject
     public float distanceFromPlayer;
     public float AnimationSpeed;
     public float GunAttackSpeed;
-    public float GunMagazineSize;
+    public int GunMagazineSize;
     public float GunBulletSpeed;
     public float GunBulletLifeTime;
     public float GunMagazineReloadTime;
